# openapi.yaml
openapi: 3.0.1
info:
  title: YWR Intelligence API
  version: "1.0.0"
  description: >
    API to retrieve YWR factor scores, QARV scores, and AI-generated investment reports for tickers.
servers:
  - url: https://api.ywr-intelligence.world
paths:
  /v1/health:
    get:
      summary: Service health
      operationId: getHealth
      tags:
        - health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /v1/factor_scores/{ticker}:
    get:
      summary: Get latest YWR factor scores for a ticker
      operationId: getFactorScores
      tags:
        - factor_scores
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
          description: >
            Returns the latest YWR factor model scores for a single stock.
            The `ticker` path parameter can be a FactSet ticker (e.g. FDS-US)
            or an alternate symbol such as an FMP-style ticker (e.g. AAPL).
            The API will automatically resolve and map the ticker internally.
            If you are unsure of the correct ticker, first call /v1/tickers/resolve.
      responses:
        '200':
          description: Latest factor scores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FactorScoreResponse'
        '404':
          description: Ticker not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/tickers/resolve:
    get:
      summary: Resolve tickers by name or alternate symbols
      operationId: resolveTicker
      tags:
        - tickers
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
          description: Company name, FactSet ticker, or alternate ticker (e.g. AAPL, MSFT, TSLA).
            Use this endpoint whenever you are not sure which ticker to use for other endpoints,
            or when the user gives a company name instead of a specific ticker.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
          description: Maximum number of matches to return
      responses:
        '200':
          description: Matching tickers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TickerResolutionResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No matching tickers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/qarv_scores/{ticker}:
    get:
      summary: Get latest QARV scores for a ticker
      operationId: getQarvScores
      tags:
        - qarv_scores
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Latest QARV scores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QarvScoreResponse'
        '404':
          description: Ticker not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/reports/{ticker}:
    get:
      summary: Get AI-generated investment report for a ticker
      operationId: getReport
      tags:
        - reports
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Generated report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '404':
          description: Ticker not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    ScoreMetadata:
      type: object
      properties:
        ticker:
          type: string
        name:
          type: string
        country:
          type: string
        industry:
          type: string
        mkt_val:
          type: number
          format: float

    FactorScoreResponse:
      allOf:
        - $ref: '#/components/schemas/ScoreMetadata'
        - type: object
          properties:
            estimate_score:
              type: number
              format: float
            value_score:
              type: number
              format: float
            price_score:
              type: number
              format: float
            total_score:
              type: number
              format: float
            date:
              type: string
              format: date

    QarvScoreResponse:
      allOf:
        - $ref: '#/components/schemas/ScoreMetadata'
        - type: object
          properties:
            qarv_metric:
              type: number
              format: float
            components:
              type: object
              additionalProperties:
                type: number
                format: float

    ReportSection:
      type: object
      properties:
        title:
          type: string
        content:
          type: string

    ReportResponse:
      type: object
      properties:
        ticker:
          type: string
        generated_at:
          type: string
          format: date-time
        summary:
          type: string
        sections:
          type: array
          items:
            $ref: '#/components/schemas/ReportSection'

    TickerMatch:
      type: object
      properties:
        factset_ticker:
          type: string
          nullable: true
        fmp_ticker:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        source:
          type: string
          description: Data source used for the match (e.g. factor_scores or ticker_map)

    TickerResolutionResponse:
      type: object
      properties:
        query:
          type: string
        matches:
          type: array
          items:
            $ref: '#/components/schemas/TickerMatch'

    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message

tags:
  - name: health
    description: Health and status endpoints
  - name: factor_scores
    description: YWR factor model scores endpoints
  - name: qarv_scores
    description: QARV score endpoints
  - name: tickers
    description: Utilities for mapping and resolving tickers
  - name: reports
    description: AI-generated investment reports